+++ src/routes/mod.rs
@@ -1,4 +1,5 @@
 pub mod follow_route;
+pub mod models;
 pub mod post_route;
 pub mod task_route;
 pub mod user_route;